fix(votes): render default tab server-side as Suspense fallback (restore archive SSR)
Reading the type tab via useSearchParams under Suspense served an empty fallback
during SSR, so the date-archive vote list became client-only (bad for SEO).
Extract a pure DailyVotesContent and use it as the fallback (default votes tab)
so the list is in the SSR HTML; the client only swaps the tab on ?type=.
